move Mock Server into it's own thing
