+++ src/client/mock_client.rs
@@ -1,11 +1,12 @@
-use futures::future::{err as FutErr, ok as FutOk};
+use futures::future::err as FutErr;
 use futures::future::{Future, IntoFuture};
 use url::Url;
 
-use errors::{Error, Result};
-use client::params::{UrlQueryParams, ParamValue};
+use errors::Result;
+use client::params::UrlQueryParams;
 use client::OauthClient;
-use client::{AccessType, AsyncPacker, ClientStorage, FutResult, ValidReq};
+use client::mock_server::*;
+use client::*;
 use client::storage::{MockMemoryStorage, MockStorageKey};
 use client::authenticator::BaseAuthenticator;
 
@@ -26,11 +27,12 @@ pub struct MockResp {
     pub body: String,
 }
 
-impl<T> From<T> for MockResp where T: Into<String> {
+impl<T> From<T> for MockResp
+where
+    T: Into<String>,
+{
     fn from(v: T) -> MockResp {
-        MockResp {
-            body: v.into(),
-        }
+        MockResp { body: v.into() }
     }
 }
 
@@ -60,7 +62,7 @@ impl MockClient {
             redirect_uri: "https://localhost/auth",
             access_type: AccessType::Grant,
             code: None,
-            server: MockServer::new()
+            server: MockServer::new(),
         })
     }
 
@@ -131,10 +133,11 @@ impl OauthClient<MockMemoryStorage> for MockClient {
     }
 
     fn request_token(&self) -> FutResult<MockResp> {
-        let token_with_params: Url = match Url::parse_with_params(self.auth.get_token_uri(), vec![("foo", "bar")]) {
-            Ok(k) => k,
-            Err(e) => return FutErr(e.into()).pack(),
-        };
+        let token_with_params: Url =
+            match Url::parse_with_params(self.auth.get_token_uri(), vec![("foo", "bar")]) {
+                Ok(k) => k,
+                Err(e) => return FutErr(e.into()).pack(),
+            };
 
         self.server.request(MockReq {
             url: token_with_params,
@@ -144,82 +147,6 @@ impl OauthClient<MockMemoryStorage> for MockClient {
 
     fn handle_token_response(self, _: MockResp, _: &mut MockMemoryStorage) -> FutResult<Self> {
         unimplemented!()
-    }
-}
-
-
-#[derive(Clone, Debug, PartialEq)]
-pub enum MockErrKind {
-    InvalidRequest,
-    UnauthorizedClient,
-    AccessDenied,
-    UnsupportedResponseType,
-    InvalidScope,
-    ServerError,
-    TemporarilyUnavailable,
-    Unknown,
-}
-
-#[derive(Clone, Debug, PartialEq)]
-pub struct MockError {
-    pub kind: MockErrKind,
-    pub description: Option<String>,
-    pub url: Option<String>,
-}
-
-#[derive(Debug, Clone, PartialEq)]
-pub struct MockServer {
-    pub error: Option<MockError>,
-}
-
-impl MockServer {
-    pub fn new() -> MockServer {
-        MockServer { error: None }
-    }
-
-    pub fn with_error(error: MockError) -> MockServer {
-        MockServer { error: Some(error) }
-    }
-
-    pub fn redirect(&self, req: MockReq) -> FutResult<MockReq> {
-        match self.error {
-            Some(ref e) => {
-                let error_kind = e.kind.clone();
-                match error_kind {
-                    _ => unimplemented!(),
-                }
-            }
-            None => match req.url.path() {
-                "/auth" => {
-                    let state = match UrlQueryParams::from(req.url.query_pairs())
-                        .get("state")
-                        .unwrap_or(ParamValue::from(""))
-                        .single()
-                        {
-                            Some(v) => v.clone(),
-                            None => String::from(""),
-                        };
-                    FutOk(MockReq {
-                        url: Url::parse_with_params(
-                            "https://localhost/example/auth",
-                            vec![("state", state), ("code", "MOCK_CODE".into())],
-                        ).unwrap(),
-                        body: String::from(""),
-                    }).pack()
-                }
-                _ => FutErr(Error::msg("404 Route not found")).pack(),
-            },
-        }
-    }
-
-    pub fn request(&self, req: MockReq) -> FutResult<MockResp> {
-        match self.error {
-            Some(_) => unimplemented!(),
-            None => match req.url.path() {
-                "/token" => FutOk(MockResp::from("{\"access_token\":\"2YotnFZFEjr1zCsicMWpAA\"}")).pack(),
-                _ => FutErr(Error::msg("404 Route not found")).pack(),
-            }
-        }
     }
 }
